Fixing various win32 errors that occur when trying to build
- Soapy.hpp needs a static_cast<long int> to build.
- bm_Buffer.cpp needs to conditionally exclude _WIN32 for
  setCpuAffinity even though we have the required headers.  We also need
  a UL to UZ type change.
- HistoryBuffer.cpp needs a UL to UZ type change.
- Scheduler.hpp need to unset ERROR defined in windows.h so the macro is
  not substituted in the work::status::ERROR method.
- MemoryMonitor.cpp windows.h needs to be included before psapi.h.
  psapi.h has alot of windows specific types that we don't know about
  before including windows.h
- thread_pool.hpp replace unsigned long type with std::size_t.  This
  squares with the type for numThreads() and will allow ULL under
  windows and UL on linux.

+++ blocks/soapy/include/gnuradio-4.0/soapy/Soapy.hpp
@@ -145,7 +145,7 @@ This block supports multiple output ports and was tested against the 'rtlsdr' an
             for (std::size_t i = 0UZ; i < rx_channels->size(); ++i) {
                 output[i] = std::span<T>(outputs[i]).subspan(0, maxSamples);
             }
-            ret = _rxStream.readStreamIntoBufferList(flags, time_ns, max_time_out_us, output);
+            ret = _rxStream.readStreamIntoBufferList(flags, time_ns, static_cast<long int>(max_time_out_us), output);
         }
         // for detailed debugging: detail::printSoapyReturnDebugInfo(ret, flags, time_ns);
